Fix multiple Bugs with high concurrency in RM　module

+++ linkis-computation-governance/linkis-manager/linkis-manager-persistence/src/main/java/org/apache/linkis/manager/persistence/impl/DefaultNodeManagerPersistence.java
@@ -74,7 +74,7 @@ public class DefaultNodeManagerPersistence implements NodeManagerPersistence {
         try {
             nodeManagerMapper.addNodeInstance(persistenceNode);
         } catch (DuplicateKeyException e) {
-            NodeInstanceDuplicateException nodeInstanceDuplicateException = new NodeInstanceDuplicateException(41001, "the node instance is exist (Node实例已存在)");
+            NodeInstanceDuplicateException nodeInstanceDuplicateException = new NodeInstanceDuplicateException(41001, "Node实例已存在");
             nodeInstanceDuplicateException.initCause(e);
             throw nodeInstanceDuplicateException;
         }
@@ -98,7 +98,7 @@ public class DefaultNodeManagerPersistence implements NodeManagerPersistence {
         } catch (DuplicateKeyException e) {
             throw new LinkisRetryException(41003, "engine instance name is exist, request of created engine will be retry");
         } catch (Exception e) {
-            NodeInstanceNotFoundException nodeInstanceNotFoundException = new NodeInstanceNotFoundException(41002, "the node instance is not  exist (Node实例不存在)");
+            NodeInstanceNotFoundException nodeInstanceNotFoundException = new NodeInstanceNotFoundException(41002, "Node实例不存在");
             nodeInstanceNotFoundException.initCause(e);
             throw nodeInstanceNotFoundException;
         }
@@ -110,7 +110,7 @@ public class DefaultNodeManagerPersistence implements NodeManagerPersistence {
         try {
             nodeManagerMapper.removeNodeInstance(instance);
         } catch (Exception e) {
-            NodeInstanceNotFoundException nodeInstanceNotFoundException = new NodeInstanceNotFoundException(41002, "the node instance is not  exist (Node实例不存在)");
+            NodeInstanceNotFoundException nodeInstanceNotFoundException = new NodeInstanceNotFoundException(41002, "Node实例不存在");
             nodeInstanceNotFoundException.initCause(e);
             throw nodeInstanceNotFoundException;
         }
@@ -151,6 +151,7 @@ public class DefaultNodeManagerPersistence implements NodeManagerPersistence {
                 persistenceNodeEntity.setMark(persistenceNode.getMark());
                 persistenceNodeEntity.setOwner(persistenceNode.getOwner());
                 persistenceNodeEntity.setStartTime(persistenceNode.getCreateTime());
+                persistenceNodeEntity.setUpdateTime(persistenceNode.getUpdateTime());
                 persistenceNodeEntitys.add(persistenceNodeEntity);
             }
         }
@@ -195,6 +196,9 @@ public class DefaultNodeManagerPersistence implements NodeManagerPersistence {
         addNodeInstance(engineNode);
         //插入关联关系，todo 异常后续统一处理
         String engineNodeInstance = engineNode.getServiceInstance().getInstance();
+        if (null == engineNode.getEMNode()) {
+            throw new PersistenceErrorException(410002, " The emNode  is null " + engineNode.getServiceInstance());
+        }
         String emNodeInstance = engineNode.getEMNode().getServiceInstance().getInstance();
         nodeManagerMapper.addEngineNode(engineNodeInstance,emNodeInstance);
     }
@@ -202,9 +206,11 @@ public class DefaultNodeManagerPersistence implements NodeManagerPersistence {
     @Override
     public void deleteEngineNode(EngineNode engineNode) throws PersistenceErrorException {
         String engineNodeInstance = engineNode.getServiceInstance().getInstance();
-        String emNodeInstance = engineNode.getEMNode().getServiceInstance().getInstance();
-        //清理 engine和em 的关系表
-        nodeManagerMapper.deleteEngineNode(engineNodeInstance,emNodeInstance);
+        if (null != engineNode.getEMNode()) {
+            String emNodeInstance = engineNode.getEMNode().getServiceInstance().getInstance();
+            //清理 engine和em 的关系表
+            nodeManagerMapper.deleteEngineNode(engineNodeInstance,emNodeInstance);
+        }
         //清理 metric信息
         metricManagerMapper.deleteNodeMetricsByInstance(engineNodeInstance);
         //metricManagerMapper.deleteNodeMetrics(emNodeId);
@@ -215,26 +221,29 @@ public class DefaultNodeManagerPersistence implements NodeManagerPersistence {
     @Override
     public EngineNode getEngineNode(ServiceInstance serviceInstance) throws PersistenceErrorException {
         //给定引擎的 serviceinstance 查到 emNode
-        PersistenceNode emNode = nodeManagerMapper.getEMNodeInstanceByEngineNode(serviceInstance.getInstance());
-        if (emNode == null) return null;
-        String emInstance = emNode.getInstance();
-        String emName = emNode.getName();
-        ServiceInstance emServiceInstance = new ServiceInstance();
-        emServiceInstance.setApplicationName(emName);
-        emServiceInstance.setInstance(emInstance);
-        AMEMNode amemNode = new AMEMNode();
-        amemNode.setMark(emNode.getMark());
-        amemNode.setOwner(emNode.getOwner());
-        amemNode.setServiceInstance(emServiceInstance);
-        amemNode.setStartTime(emNode.getCreateTime());
-
         AMEngineNode amEngineNode = new AMEngineNode();
         amEngineNode.setServiceInstance(serviceInstance);
         PersistenceNode engineNode = nodeManagerMapper.getNodeInstance(serviceInstance.getInstance());
+        if(null == engineNode) {
+            return null;
+        }
         amEngineNode.setOwner(engineNode.getOwner());
         amEngineNode.setMark(engineNode.getMark());
         amEngineNode.setStartTime(engineNode.getCreateTime());
-        amEngineNode.setEMNode(amemNode);
+        PersistenceNode emNode = nodeManagerMapper.getEMNodeInstanceByEngineNode(serviceInstance.getInstance());
+        if (emNode != null) {
+            String emInstance = emNode.getInstance();
+            String emName = emNode.getName();
+            ServiceInstance emServiceInstance = new ServiceInstance();
+            emServiceInstance.setApplicationName(emName);
+            emServiceInstance.setInstance(emInstance);
+            AMEMNode amemNode = new AMEMNode();
+            amemNode.setMark(emNode.getMark());
+            amemNode.setOwner(emNode.getOwner());
+            amemNode.setServiceInstance(emServiceInstance);
+            amemNode.setStartTime(emNode.getCreateTime());
+            amEngineNode.setEMNode(amemNode);
+        }
         return amEngineNode;
     }
 
@@ -242,7 +251,9 @@ public class DefaultNodeManagerPersistence implements NodeManagerPersistence {
     public List<EngineNode> getEngineNodeByEM(ServiceInstance serviceInstance) throws PersistenceErrorException {
         //给定EM的 serviceinstance
         PersistenceNode emNode = nodeManagerMapper.getNodeInstance(serviceInstance.getInstance());
-
+        if (null == emNode) {
+            return new ArrayList<>();
+        }
         List<PersistenceNode> engineNodeList = nodeManagerMapper.getNodeInstances(serviceInstance.getInstance());
         List<EngineNode> amEngineNodeList =new ArrayList<>();
         for (PersistenceNode engineNode : engineNodeList){
@@ -266,4 +277,5 @@ public class DefaultNodeManagerPersistence implements NodeManagerPersistence {
         }
         return amEngineNodeList;
     }
+
 }
